refactor(tests): use component barrels instead of deep imports
Tests outside a component must use the component's barrel per the
import discipline convention. 10 sites corrected:

- helpers/TemplateFixtures.ts: CURRENT_SCHEMA_VERSION via
  charging-station/index.js (was TemplateMigrations.js)
- 3 ui-server tests: logger via utils/index.js (was Logger.js)
- ui-server/UIMCPServer.integration.test.ts: Configuration via
  utils/index.js (was Configuration.js)
- 3 ocpp/2.0 tests: buildConfigKey/addConfigurationKey/setConfiguration\
  KeyValue/deleteConfigurationKey/getConfigurationKey via
  charging-station/index.js (was ConfigurationKeyUtils.js)
- OCPP20ServiceUtils-TransactionEvent.test.ts: merge ConnectorStatus
  and EmptyObject into a single type-import from types/index.js
  (was ConnectorStatus.js direct); addConfigurationKey via
  charging-station/index.js.

All 2954/2960 tests still pass. Deep imports remain only where the
target symbol is a genuine barrel gap (ui-server/, ocpp sub-components,
performance/storage/) — those are follow-up work.
