+++ tests/charging-station/ui-server/UIMCPServer.integration.test.ts
@@ -15,7 +15,7 @@ import { afterEach, beforeEach, describe, it } from 'node:test'
 import { UIMCPServer } from '../../../src/charging-station/ui-server/UIMCPServer.js'
 import { HttpMethod } from '../../../src/charging-station/ui-server/UIServerUtils.js'
 import { ApplicationProtocol, ConfigurationSection } from '../../../src/types/index.js'
-import { Configuration } from '../../../src/utils/Configuration.js'
+import { Configuration } from '../../../src/utils/index.js'
 import { standardCleanup } from '../../helpers/TestLifecycleHelpers.js'
 import { createMockBootstrap, createMockUIServerConfiguration } from './UIServerTestUtils.js'
