+++ tests/charging-station/ui-server/UIWebSocketServer.test.ts
@@ -25,7 +25,7 @@ import {
   ProcedureName,
   ResponseStatus,
 } from '../../../src/types/index.js'
-import { logger } from '../../../src/utils/Logger.js'
+import { logger } from '../../../src/utils/index.js'
 import { createLoggerMocks, standardCleanup } from '../../helpers/TestLifecycleHelpers.js'
 import { TEST_UUID } from './UIServerTestConstants.js'
 import {
